modify api

+++ routes/api.js
@@ -26,7 +26,11 @@ router.get('/notes', function (req, res, next) {
 
 /* 添加 */
 router.post('/notes/add', function (req, res, next) {
+  if(!req.session || !req.session.user){
+    return res.send({status: 0, errorMsg: '请先登录'})
+  }
   var note = req.body.note;
+  var userid=req.session.user.id;
   if (!note) {
     return res.send({
       status: 0,
@@ -35,7 +39,8 @@ router.post('/notes/add', function (req, res, next) {
   }
 
   Note.create({
-    text: note
+    text: note,
+    userid:userid
   }).then(function (data) {
     res.send({
       status: 1
@@ -52,8 +57,12 @@ router.post('/notes/add', function (req, res, next) {
 /* 编辑 */
 
 router.post('/notes/edit', function (req, res, next) {
+  if(!req.session || !req.session.user){
+    return res.send({status: 0, errorMsg: '请先登录'})
+  }
   var editid = req.body.id;
   var note = req.body.note;
+  var userid=req.session.user.id;
   if (!note) {
     return res.send({
       status: 0,
@@ -64,7 +73,8 @@ router.post('/notes/edit', function (req, res, next) {
     text: note
   }, {
     where: {
-      id: editid
+      id: editid,
+      userid:userid
     }
   }).then(function (data) {
     console.log(data);
@@ -81,11 +91,15 @@ router.post('/notes/edit', function (req, res, next) {
 
 /* 删除 */
 router.post('/notes/delete', function (req, res, next) {
+  if(!req.session || !req.session.user){
+    return res.send({status: 0, errorMsg: '请先登录'})
+  }
   var deleteid = req.body.id;
-  console.log(deleteid);
+  var userid=req.session.user.id;
   Note.destroy({
     where: {
-      id: deleteid
+      id: deleteid,
+      userid:userid
     }
   }).then(function (data) {
     if (data !== 0) {
